some changes

+++ ejbModule/remarema/services/network/NetworkServiceBean.java
@@ -7,6 +7,7 @@ import javax.persistence.*;
 
 import remarema.api.*;
 import remarema.domain.Network;
+import remarema.domain.Node;
 
 /**
  * Session Bean implementation class NetworkServiceBean
@@ -64,8 +65,12 @@ public class NetworkServiceBean {
 			
 			detail.setNetworkID(result.getNetworkID());
 			detail.setNetworkName(result.getNetworkName());
-			detail.setNetworkParentID(result.getNetworkID());
-			//detail.setNetworkParentName(result.getParent().getNetworkName());
+			
+			if(result.hasParentNetwork()){
+				Network parent = result.getParent();
+				detail.setNetworkParentID(parent.getNetworkID());
+				detail.setNetworkParentName(parent.getNetworkName());
+			}
 			
 			networkDetail.add(detail);
 		
@@ -82,19 +87,54 @@ public class NetworkServiceBean {
 	}
 	
 	
-	public void networkUpdate(NetworkDetail parameterObject) {
-		Network nw = em.find(Network.class, parameterObject.getNetworkID());
-		em.getTransaction().begin();
-		nw.setNetworkName(parameterObject.getNetworkName());
-		em.getTransaction().commit();
+	public void networkUpdate(UpdateNetwork command) {
+		Network nw = em.find(Network.class, command.getNetworkID());
+		nw.setNetworkName(command.getNetworkName());
+		Network parentNetwork = findParentNetwork(command.getNetworkParentName());
+		parentNetwork.setNetworkName(command.getNetworkName());
+		nw.setParent(parentNetwork);
+		em.flush();
 	}
 	
-
-	public void removeNetwork(NetworkDetail parameterObject) {
+	
+	public void removeNetwork(NetworkDetail parameterObject) throws ChildNotEmptyException {
 		Network nw = em.find(Network.class, parameterObject.getNetworkID());
+		
 		if (nw != null) {
-			em.remove(nw);
+			if(nw.getChildren().isEmpty()){
+				em.remove(nw);
+			}
+			else{
+				throw new ChildNotEmptyException("Dieses Netzwerk beinhaltet Clients!");
+			}
 		}
+		else{
+			throw new IllegalArgumentException("Netzwerk ist null");
+		}
+	}
+
+	public NetworkDetail getNetworkDetailForNetworkID(NetworkDetail networkDetail) {
+		int networkID = networkDetail.getNetworkID();
+		return mapNetworkToNetworkDetail(loadNetwork(networkID));
+	}
+	
+	NetworkDetail mapNetworkToNetworkDetail(Network network){
+		NetworkDetail nwd = new NetworkDetail();
+		nwd.setNetworkID(network.getNetworkID());
+		nwd.setNetworkName(network.getNetworkName());
+//		
+//		Network networkParent = network.getParent();
+//		if(networkParent != null){
+//			nwd.setNetworkParentID(networkParent.getNetworkID());
+//			nwd.setNetworkParentName(networkParent.getNetworkName());
+//		}
+		return nwd;
+	}
+
+	Network loadNetwork(int networkID) {
+		TypedQuery<Network> query = em.createQuery(
+				"SELECT o FROM Network o WHERE o.networkID =" + networkID, Network.class);
+		return query.getSingleResult();
 	}
 
 }
